refactor(agents,config): split headless + settings by concern under the LOC cap (#2565)
Merging #2877 (model tiering) into the #2565 harness seam pushed two core
files past the 500 effective-LOC module-health cap. Split each by concern,
behaviour-preserving — no suppression, no cap bump, no noqa.

agents/headless.py (508 -> 431): the SDK option-building concern moves to a
new sibling agents/_headless_options.py — _build_options plus its
model-tiering glue, _resolve_task_cwd, _get_resume_session_id, and the spawn
constants (_PERMISSION_MODE, _MAX_TURNS, _DISALLOWED_TOOLS, _ADAPTIVE_THINKING,
UUID_RE). headless keeps run_headless, the drive/heartbeat loop, the watchdog/
budget/outcome types, and the result recorders. run_headless imports
_build_options for use; the other moved names' consumers (teams.pane_spawn,
core.management.commands.tasks, and the tests) now import from the true home.

config/settings.py (503 -> 327): the field-value coercers move to a new
config/setting_parsers.py (the _parse_strict_* / _parse_env_* / list coercers
and the _default_handover_mirror_path helper). The override registries and the
dataclasses stay in settings, which imports the coercers to build them — so
every `from teatree.config.settings import _parse_*` site keeps resolving. The
teatree.config facade re-exports the coercers from setting_parsers
(_parse_disk_cache_allowlist repointed there; all others still surface via
settings' registry use).

Both new modules are all-private-function (0 public module functions) and well
under the cap (91 and 204 effective LOC). No behaviour change.

Open questions & assumptions:
- Moved names' external consumers were repointed to _headless_options rather
  than re-exported through headless: ruff PLC0414 rejects a redundant-alias
  re-export in a non-__init__ module, and an __all__ of private names would
  misrepresent the module surface. _build_options stays importable from
  headless because run_headless genuinely uses it.
- The settings override registries were kept in settings.py (tied to the
  dataclass field names) rather than moved with the coercers; the coercers are
  the cohesive, reusable cluster.
